[CHORE] #53 : imageProcess, objectDetect -> textDetection, objectDetection rename
