+++ client/app/src/main/res/layout/fragment_object_detection.xml
@@ -2,11 +2,11 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.camera.object_detect.SearchCategoryFragment">
+    tools:context=".presentation.object_detection.ObjectDetectionFragment">
     <data>
         <variable
             name="viewModel"
-            type="com.example.healthc.presentation.camera.object_detect.SearchCategoryViewModel"
+            type="com.example.healthc.presentation.object_detection.ObjectDetectionViewModel"
             />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
