+++ client/app/src/main/res/layout/fragment_text_detection.xml
@@ -3,11 +3,11 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.camera.image_process.ImageProcessFragment">
+    tools:context=".presentation.text_detection.TextDetectionFragment">
     <data>
         <variable
             name="viewModel"
-            type="com.example.healthc.presentation.camera.image_process.ImageProcessViewModel"
+            type="com.example.healthc.presentation.text_detection.TextDetectionViewModel"
             />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
